+++ CHANGELOG.md
@@ -1,5 +1,7 @@
-# DEV
+# 0.13.0
 
+* removed grunticon and replaced it with svgcss #8
+  * dropped PNG-fallback support!
 * set `attr-value-double-quotes` and `attr-value-not-empty` to false and added `space-tab-mixed-disabled` in htmlhint options
 * pinned handlebars-helper-partial in package.json to version 0.1.2 (there are conficts with 0.2.0)